+++ frontend/src/App.jsx
@@ -6,7 +6,6 @@ import SignIn from "./pages/signin";
 import JobSeekerDashboard from "./pages/JobSeeker/DashBoard";
 import CompanyRepDashboard from "./pages/CompanyRep/DashBoard";
 import JobFinder from "./pages/JobSeeker/JobFinder";
-import EmployerJobPosting from "./pages/CompanyRep/EmployerJobPosting";
 import Landing from "./pages/Landing";
 import Jobposts from "./pages/CompanyRep/Jobposts";
 
@@ -20,7 +19,6 @@ function App() {
       
       <Route path="/jobseeker/dashboard" element={<JobSeekerDashboard />} />
       <Route path="/jobs" element={<JobFinder />} />
-      <Route path="/employer" element={<EmployerJobPosting />} />
       
       <Route path="/companyrep/jobposts" element={<Jobposts />} />
